// src/interfaces/clima.interface.ts
export type IClima = {
  coord: { lon: -47.6054; lat: -21.2067 };
  weather: [{ id: 804; main: "Clouds"; description: "nublado"; icon: "04d" }];
  base: "stations";
  main: {
    temp: 25.79;
    feels_like: 26.33;
    temp_min: 25.79;
    temp_max: 25.79;
    pressure: 1013;
    humidity: 73;
  };
  visibility: 10000;
  wind: { speed: 2.57; deg: 270 };
  clouds: { all: 100 };
  dt: 1643743427;
  sys: {
    type: 1;
    id: 8428;
    country: "BR";
    sunrise: 1643705569;
    sunset: 1643752505;
  };
  timezone: -10800;
  id: 3447720;
  name: "Serrana";
  cod: 200;
};